Codec.Console.Options

Description

Command-line argument decoder.

Decode

data Decoder

Argument decoder.

Constructors

Short
Short option group.

Fields

!Char

Option name.

(Maybe OsString)

Option argument as remainder of the group.

ShortDecoder

Next short option in the group.

Long
Long option.

Fields

!ShortText

Option name.

!(Maybe OsString)

Option argument.

Argument
Non-option argument.

End
End of options (--).

Fields

!(Maybe OsString)

--=ARG is still a possible (albeit invalid) sequence.

Undecodable
Argument begins with a hyphen, but some characters could not be decoded.

data ShortDecoder

Short option group decoder.

Constructors

Short'
Next short option.

Fields

!Char

Option name.

(Maybe OsString)

Option argument as remainder of the group.

ShortDecoder

Next short option in the group.

End'
End of group.

Undecodable'
Short option character could not be decoded.

decode

Arguments

:: Bool
True for Windows, False for POSIX.

-> ByteArray
Unwrapped WindowsString/PosixString.

-> ShortByteString
WTF-8 byte array (same encoding as FilePath).

-> Decoder

Interpret a single argument.

Decoded option names are guaranteed to only contain characters in the POSIX Portable Character Set.

Requires the same argument in two forms.