Data.Console.Option

Description

Safe haven for the Flavor type since it's needed for downstream printing, but wouldn't be convenient to have in the core user-facing modules.

Documentation

data Name

Option name.

Constructors

Short !Char
As in -s.

Long !Text
As in --long-option.

data Option m help f

Single option.

Constructors

Option

Fields

!(NonEmpty Name)

Names which can be used to invoke the option.

!(Flavor m help f)

data Flavor m help f

Option behavior coupled with a function the option invokes.

Constructors

Plain (m f)

Optional help (m (Maybe String -> f))

Required help (m (String -> f))