Data.Console.Options

Description

Options is a spine-strict tree of command-line options.

Laziness

Evaluating the root of the tree (i.e. (_ :: Options m f)) to weak head normal form evaluates the entire spine of the tree to normal form.

Itself

type OptionsQ f = Q (Options (Code Q :: Type -> Type) f)

Convenience synonym.

data Options (m :: Type -> Type) f

Spine-strict option tree.

Construct

empty :: forall (m :: Type -> Type) f. Options m f

Empty option tree.

Insert

data Name

Option name.

Constructors

Short !Char
As in -s.

Long !ShortText
As in --long-option.

data Flavor help f arg where

Option behavior.

Constructors

Plain :: forall help f. Flavor help f f
Option takes no arguments.

Optional
Option takes an optional argument.

Fields

:: forall help f. help

Help description for the argument, i.e. ARG in -o, --option[=ARG].

-> Flavor help f (Maybe OsString -> f)

Required
Option requires an argument.

Fields

:: forall help f. help

Help description for the argument, i.e. ARG in -o, --option[=ARG].

-> Flavor help f (OsString -> f)

data Option help f arg

Single option.

Constructors

Option

Fields

[Name]

Names that can be used to invoke the option.

(Flavor help f arg)

insert :: Option help f arg -> arg -> Options Identity f -> Options Identity f

Insert an option into the tree.

If the option has an empty list of names, nothing is added to the tree. Empty long names are similarly treated as no-ops. When inserting duplicate names new values take precedence.

insertQ :: Option help f arg -> Code Q arg -> OptionsQ f -> OptionsQ f

Insert an option into the tree at compilation time.

Emits warnings if any of the provided option names are inaccessible, overwrite previously added names, or contain non-portable characters. Operates identically to insert otherwise.

Precompile

sequenceCode :: OptionsQ f -> Code Q (Options Identity f)

Precompile the option tree.

Analyze

suggestions

Arguments

:: Options Identity f

-> ShortByteString
Full argument as a WTF-8 byte array.

-> Name
Option name in question.

-> [Name]

Fuzzy search for Unrecognized options.