System.Console.Options

Description

Command-line option parsing.

Implementation

Description of the argument syntax can be found at GNU Program Argument Syntax Conventions page, which extends the short-option-only POSIX Utility Conventions.

On POSIX-compliant systems string comparisons rely on POSIX portable characters always being represented in a single byte, per POSIX Character Set specification.

Parse

data Failure

Argument parsing error.

Constructors

NonPortable
Option contains characters outside of the POSIX Portable Character Set.

Fields

!OsString

Raw argument.

!ShortByteString

Argument converted to WTF-8.

Unrecognized
Option name isn't on the list.

Fields

!OsString

Raw argument.

!ShortByteString

Argument converted to WTF-8.

!Name

Parsed option name.

Unsaturated
Option requires an argument, but none was provided.

Fields

!Name

Parsed option name.

Oversaturated
Option takes no arguments, but was provided with one.

Fields

!ShortText

Long option name, may be empty (consider --=ARG).

!OsString

Argument passed to the option.

data Outcome r

Parsing outcome.

Constructors

Success [OsString]
Reached end of options successfully.

Returns modified state and the list of non-option arguments in the order they were encountered.

Broken r
Optional early exit.

Failure !Failure

parse :: Options Identity (s -> IO (Either r ())) -> s -> [OsString] -> IO (Outcome r)

Parse command-line options.

Treats characters outside of POSIX Portable Character Set as incomparable.

Relies on getFileSystemEncoding on POSIX-compliant platforms.