Codec.Console.Options

Description

Decoder definition and several predefined runners.

Documentation

Decoder

newtype Decoder f

Decoding algorithm without any input applied.

Constructors

Decoder ([String] -> Step f)

data Step f

Single decoding step, corresponding to processing 1-2 arguments.

Constructors

Single
Parsed a standalone option.

Fields

f

Option itself.

[String]

Rest of the argument list.

Multiple
Parsed a short option that is part of a group.

Fields

f

Option itself.

(Step f)

Next option in this group.

Argument
Parsed an argument.

Fields

String

Argument itself.

[String]

Rest of the argument list.

Done
Reached end of options successfully.

Fields

[String]

Remaining non-option arguments, if any.

Failed !Failure
Could not parse the argument.

Runner

data Name

Option name.

Constructors

Short !Char
As in -s.

Long !Text
As in --long-option.

data Failure

Parsing error.

Constructors

Unrecognized
Option name isn't on the list.

Fields

!String

The argument that this option is a part of, without the preceding dash(es).

!Name

The option name in question.

Unsaturated !Name
Option requires an argument, but none was provided.

Oversaturated
Option takes no arguments, but was provided with one.

Fields

!Text

Long option name, may be empty (consider --=ARG).

!String

Argument passed to the option.

data Result r s

General result type.

Constructors

Success s [String]
Reached end of options successfully.

Returns modified state and the list of non-option arguments in the order they were encountered.

Break r
Optional early exit.

Failure !Failure

run :: Decoder (s -> Either r s) -> s -> [String] -> Result r s

Run the decoder over pure state, sequencing operations left-to-right.

runM :: Monad m => Decoder (s -> m (Either r s)) -> s -> [String] -> m (Result r s)

Run the decoder over pure state, sequencing monadic operations left-to-right.

runInOrder

Arguments

:: (String -> s -> Either r s)
Function to apply to every argument.

-> Decoder (s -> Either r s)

-> s

-> [String]

-> Result r s

Run the decoder over pure state, applying the given function to every argument in place and sequencing operations left-to-right.