System.Console.Options

Description

Runtime-evaluated command-line option parsing.

Implementation

Description of the argument syntax can be found at GNU Program Argument Syntax Conventions page, which extends the short-option-only POSIX Utility Conventions.

Documentation

data Name

Option name.

Constructors

Short !Char
As in -s.

Long !Text
As in --long-option.

Construct

data Option m help f

Single option.

Constructors

Option

Fields

!(NonEmpty Name)

Names which can be used to invoke the option.

!(Flavor m help f)

Flavor

data Flavor m help f

Option behavior coupled with a function the option invokes.

plain :: f -> Flavor Identity help f

Option takes no arguments.

optional

Arguments

:: help
Help description for the argument, i.e. ARG in -o, --option[=ARG].

-> (Maybe String -> f)

-> Flavor Identity help f

Option takes an optional argument.

required

Arguments

:: help
Help description for the argument, i.e. ARG in -o, --option=ARG.

-> (String -> f)

-> Flavor Identity help f

Option requires an argument.

Combine

data Options m f

Spine-lazy option dictionary.

none :: Options Identity f

Empty option dictionary.

(.>) :: Options Identity f -> Option Identity help f -> Options Identity f infixl 5

Insert an option into the dictionary.

If the option has an empty list of names, nothing is added to the dictionary. Empty long names are similarly treated as no-ops. When inserting duplicate names new values take precedence.

Precompile

Refer to System.Console.Options.TH on how to construct OptionsQ.

precompile :: OptionsQ f -> Code Q (Options Identity f)

Precompile the option dictionary.

Analyze

suggestions

Arguments

:: String
Argument that the option name is a part of, without the preceding dash(es).

-> Name
Option name in question.

-> Options Identity f

-> [Name]

Fuzzy search for Unrecognized options.

Parse

Decoder

data Order

Decoder operation mode. Similar to System.Console.GetOpt.ArgOrder.

Constructors

RequireOrder
First non-option marks the end of options

Permute
Non-options are freely interspersed with options

data Decoder f

Decoding algorithm without any input applied.

decode :: Order -> Options Identity f -> Decoder f

Construct a decoder.

Runner

data Failure

Parsing error.

Constructors

Unrecognized
Option name isn't on the list.

Fields

!String

The argument that this option is a part of, without the preceding dash(es).

!Name

The option name in question.

Unsaturated !Name
Option requires an argument, but none was provided.

Oversaturated
Option takes no arguments, but was provided with one.

Fields

!Text

Long option name, may be empty (consider --=ARG).

!String

Argument passed to the option.

data Result r s

General result type.

Constructors

Success s [String]
Reached end of options successfully.

Returns modified state and the list of non-option arguments in the order they were encountered.

Break r
Optional early exit.

Failure !Failure

run :: Decoder (s -> Either r s) -> s -> [String] -> Result r s

Run the decoder over pure state, sequencing operations left-to-right.

Refer to Codec.Console.Options for both the unwrapping of the Decoder and other runner functions.

Re-exports

newtype Identity a

Identity functor and monad. (a non-strict monad)

Since: base-4.8.0.0

Constructors

Identity

Fields

runIdentity :: a